requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
PyJWT[crypto]>=2.8.0
python-dotenv>=1.1.0
certifi>=2024.0.0

# MCP (client + FastMCP server; CLI extra for mcp tooling)
mcp[cli]>=1.8.0

# LLM + REPL
anthropic>=0.51.0
prompt-toolkit>=3.0.51
